% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roccurve.R
\name{auc}
\alias{auc}
\title{Returns the area under the curve value}
\usage{
auc(q, y = attr(q, "y"))
}
\arguments{
\item{q}{the predicted values}

\item{y}{a list of the actual classes of \code{q}}
}
\value{
the area under the curve value
}
\description{
Returns the area under the curve value as a fraction.
}
\examples{
df <- with(diabetes, cbind(y, X))
lm.y <- glm(y ~ ., data = df, family = binomial())
print(with(lm.y, auc(fitted.values, y)))

}
