% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispline.R
\name{isplinebasis}
\alias{isplinebasis}
\title{Transform a given data into I-splines}
\usage{
isplinebasis(x, knots, d)
}
\arguments{
\item{x}{a scalar or vector of values which will be transformed into splines}

\item{knots}{a vector of knot values of the splines}

\item{d}{the polynomial degree of the splines}
}
\value{
a matrix with for each value of x the corresponding spline values.
}
\description{
Inner function call to create I-splines based on the
user defined \code{knots}  and polynomial degree \code{d} of the splines
}
