% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge.R
\name{plot.hinge}
\alias{plot.hinge}
\title{Plot the hinge function}
\usage{
\method{plot}{hinge}(x, y = 1, z = NULL, ...)
}
\arguments{
\item{x}{The hinge object returned from \code{getHinge}.}

\item{y}{Specifies the class (\code{-1} or \code{1})
to be plotted for the hinge error.}

\item{z}{If specified, the majorization function with the supporting point
\code{z} will also be plotted.}

\item{...}{Other arguments passed to plot method.}
}
\description{
This function plots the hinge object created by \code{getHinge}.
}
\examples{
hingefunction <- getHinge()
## plot hinge function value
plot(hingefunction, z = 3)
}
