% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmmajcrossval.R
\name{plot.svmmajcrossval}
\alias{plot.svmmajcrossval}
\title{Plot the cross validation output}
\usage{
\method{plot}{svmmajcrossval}(x, type = "grid", ...)
}
\arguments{
\item{x}{the \code{svmmajcrossval} object}

\item{type}{the type of graph being shown, possible values are
\code{'grid'} for the missclassification rate per lambda value,
\code{'profile'} the distribution of predicted values
of the classes per lambda value}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Shows the results of the cross validation graphically.
Possible graphics are among others the distribution of
the predicted values \code{q} per class per lambda value
and the misclassification rate per lambda.
}
