% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsvmmaj.R
\name{plotWeights}
\alias{plotWeights}
\title{Plot the weights of all attributes from the trained SVM model}
\usage{
plotWeights(object, plotdim = c(3, 3), ...)
}
\arguments{
\item{object}{The model returned from \code{svmmaj}.}

\item{plotdim}{A vector of the form \code{c(nr, nc)}. Subsequent figures
will be drawn in an \code{nr}-by-\code{nc} array on the device.}

\item{...}{other parameters given to the \code{plot} function}
}
\description{
Shows, one graph per attribute, the weights of all
attributes. The type of graph depends on the type of the attribute: the
spline line of the corresponding attribute in case a spline has been used, a
bar plot for categorical and logical values, and a linear line for all other
type of the attribute values. This function cannot be used in a model with a
non-linear kernel.
}
