% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{predict.transDat}
\alias{predict.transDat}
\title{Perform the transformation based on predefined settings}
\usage{
\method{predict}{transDat}(
  x,
  attrib = NULL,
  values = NULL,
  standardization = NULL,
  splineInterval = NULL,
  splineDegree = NULL
)
}
\arguments{
\item{x}{a (new) vector of numerics to be transformed}

\item{attrib}{either a list of settings, or \code{NULL} in case
the attributes are given as separate input}

\item{values}{a vector of levels in case \code{x} is a factor}

\item{standardization}{the standardization rules from \code{normalize}}

\item{splineInterval}{the knots to be used for spline basis}

\item{splineDegree}{the polynomial degree of the splines}
}
\value{
a transformed data based on the user defined settings
}
\description{
Given the input parameters, which are generated from
\code{transformdata}, it performs the same transformation
with the same settings to the given input
}
