% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmmajcrossval.R
\name{print.svmmajcrossval}
\alias{print.svmmajcrossval}
\alias{summary.svmmajcrossval}
\title{Print SVMMaj cross validation results}
\usage{
\method{print}{svmmajcrossval}(x, ...)

\method{summary}{svmmajcrossval}(object, ...)
}
\arguments{
\item{x}{the cross-validation output from \code{\link{svmmajcrossval}}}

\item{...}{ignored}

\item{object}{the output object from \code{svmmajcrossval}}
}
\description{
Prints the result from the cross validation procedure in
\code{\link{svmmajcrossval}}.
}
