% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roccurve.R
\name{roccurve}
\alias{roccurve}
\title{Plot the ROC curve of the predicted values}
\usage{
roccurve(q, y = attr(q, "y"), class = 1, ...)
}
\arguments{
\item{q}{the predicted values}

\item{y}{a list of the actual classes of \code{q}}

\item{class}{the base class to show the roc-curve}

\item{...}{additional parameters given as input to the \code{plot} function}
}
\description{
Given the predicted values \code{q} and its corresponding
observed classes \code{y}, it shows its separation performances
by showing the roc-curve.
}
\examples{
model <- svmmaj(diabetes$X, diabetes$y)
roccurve(model$q)
}
