% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supermarket1996}
\alias{supermarket1996}
\title{Supermarket data 1996}
\format{This dataframe contains the following columns
\itemize{
  \item{STORE}{Identifier of the store}
  \item{CITY}{The city of the store}
  \item{ZIP}{The zip code of the store}
  \item{GROCERY_sum}{}
  \item{GROCCOUP_sum}{}
  \item{AGE9}{}
  \item{AGE60}{}
  \item{ETHNIC}{}
  \item{EDUC}{}
  \item{NOCAR}{}
  \item{INCOME}{}
  \item{INCSIGMA}{}
  \item{HSIZEAVG}{}
  \item{HSIZE1}{}
  \item{HSIZE2}{}
  \item{HSIZE34}{}
  \item{HSIZE567}{}
  \item{HH3PLUS}{}
  \item{HH4PLUS}{}
  \item{HHSINGLE}{}
  \item{HHLARGE}{}
  \item{WORKWOM}{}
  \item{SINHOUSE}{}
  \item{DENSITY}{}
  \item{HVAL150}{}
  \item{HVAL200}{}
  \item{HVALMEAN}{}
  \item{SINGLE}{}
  \item{RETIRED}{}
  \item{UNEMP}{}
  \item{WRKCH5}{}
  \item{WRKCH17}{}
  \item{NWRKCH5}{}
  \item{NWRKCH17}{}
  \item{WRKCH}{}
  \item{NWRKCH}{}
  \item{WRKWCH}{}
  \item{WRKWNCH}{}
  \item{TELEPHN}{}
  \item{MORTGAGE}{}
  \item{NWHITE}{}
  \item{POVERTY}{}
  \item{SHPCONS}{}
  \item{SHPHURR}{}
  \item{SHPAVID}{}
  \item{SHPKSTR}{}
  \item{SHPUNFT}{}
  \item{SHPBIRD}{}
  \item{SHOPINDX}{}
  \item{SHPINDX}{}
}}
\description{
This
}
\examples{
head(supermarket1996, 3)
}
\keyword{datasets}
