% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{X.svmmaj}
\alias{X.svmmaj}
\title{Returns transformed attributes}
\usage{
X.svmmaj(object, X.new, weights = NULL)
}
\arguments{
\item{object}{Model which has been trained beforehand using
\code{\link[SVMMaj]{svmmaj}}.}

\item{X.new}{Attribute matrix of the objects to be predicted, which has the
same number of attributes as the untransformed attribute matrix in
\code{model}.}

\item{weights}{The weight of observation as the relative importance of the
prediction error of the observation.}
}
\description{
For efficiency use in svmmajcrossval
}
